Adicionadas permissões para os objetos de Cenário

+++ universo_nextlevel/core/views.py
@@ -1,5 +1,6 @@
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
+from django.http.response import Http404
 from django.urls import reverse_lazy, reverse
 from django.shortcuts import get_object_or_404, redirect
 from django.views.generic import (
@@ -11,6 +12,7 @@ from django.views.generic import (
 
 from django_filters.views import FilterView
 from django_twitch_auth.views import TwitchLoginMixin
+from rules.contrib.views import PermissionRequiredMixin
 
 from universo_nextlevel.mixins import RequestUserMixin
 
@@ -45,10 +47,11 @@ class CenarioCreateView(LoginRequiredMixin, RequestUserMixin, CreateView):
         return context
 
 
-class CenarioUpdateView(LoginRequiredMixin, RequestUserMixin, UpdateView):
+class CenarioUpdateView(LoginRequiredMixin, PermissionRequiredMixin, RequestUserMixin, UpdateView):
     model = Cenario
     form_class = CenarioForm
     success_url = reverse_lazy('cenario_list')
+    permission_required = 'core.can_edit_cenario_instance'
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -64,5 +67,7 @@ class CenarioDetailView(LoginRequiredMixin, DetailView):
 @login_required
 def cenario_delete_view(request, pk):
     cenario = get_object_or_404(Cenario, pk=pk)
+    if not request.user.has_perm('core.can_delete_cenario_instance', cenario):
+        raise Http404
     cenario.delete()
     return redirect(reverse('cenario_list'))
